on mit a la norme le projet et on la bien strcuturer
